Separate some nodes into GameDisplayBox.fxml
