+++ CapsLock/src/capslock/DefaultTheme.css
@@ -1,11 +1,3 @@
 .tile-pane{
     -fx-background-color: #212425;
 }
-
-.v-box{
-    -fx-background-color: #999999;
-}
-
-.label{
-    -fx-text-fill: white;
-}
